Support absolute address for memory computations
Also adds minor code improvements
